removed gcc 4.1 compiler warnings on linux at daimi
Ignore-this: 2355b15a6e03c85278d39133aad82db8

darcs-hash:20100803152051-3a0eb-b83a8b7f71f8e4e4afc7897486631690ec0d2091.gz

+++ Resources/CairoFont.cpp
@@ -142,7 +142,8 @@ Vector<2,int> CairoFont::TextDim(string s) {
                             slant, weight);
     cairo_set_font_size (cr, ptsize);
     cairo_text_extents (cr, s.c_str(), &te);
-    Vector<2,int> dim(te.width-te.x_bearing, te.height-te.y_bearing);
+    Vector<2,int> dim((int)(te.width-te.x_bearing),
+		      (int)(te.height-te.y_bearing));
     return dim;
 }
